+++ tencentcloud/cam/v20190116/errors.go
@@ -173,6 +173,9 @@ const (
 	// PrincipalService不存在。
 	INVALIDPARAMETER_PRINCIPALSERVICENOTEXIST = "InvalidParameter.PrincipalServiceNotExist"
 
+	// 消息接收人超出限制。
+	INVALIDPARAMETER_RECEIVEROVERLIMIT = "InvalidParameter.ReceiverOverLimit"
+
 	// 策略语法中资源内容不正确。
 	INVALIDPARAMETER_RESOURCECONTENTERROR = "InvalidParameter.ResourceContentError"
 
@@ -245,6 +248,9 @@ const (
 	// 用户名不合法。
 	INVALIDPARAMETER_USERNAMEILLEGAL = "InvalidParameter.UserNameIllegal"
 
+	// 用户未实名。
+	INVALIDPARAMETER_USERNOTAUTH = "InvalidParameter.UserNotAuth"
+
 	// 用户对象不存在。
 	INVALIDPARAMETER_USERNOTEXIST = "InvalidParameter.UserNotExist"
 
@@ -268,6 +274,9 @@ const (
 
 	// 身份提供商已达到上限。
 	LIMITEXCEEDED_IDENTITYFULL = "LimitExceeded.IdentityFull"
+
+	// 请求超出限制。
+	LIMITEXCEEDED_UIN = "LimitExceeded.Uin"
 
 	// 每个账号最多支持两个AccessKey。
 	OPERATIONDENIED_ACCESSKEYOVERLIMIT = "OperationDenied.AccessKeyOverLimit"