+++ tencentcloud/ess/v20201111/models.go
@@ -56,6 +56,8 @@ type ApproverInfo struct {
 	// 1：个人
 	// 3：企业静默签署
 	// 注：类型为3（企业静默签署）时，此接口会默认完成该签署方的签署。静默签署仅进行盖章操作，不能自动签名。
+	// 7: 个人自动签署，适用于个人自动签场景。
+	// 注: 个人自动签场景为白名单功能, 使用前请联系对接的客户经理沟通。
 	ApproverType *int64 `json:"ApproverType,omitempty" name:"ApproverType"`
 
 	// 签署人的姓名
@@ -64,7 +66,9 @@ type ApproverInfo struct {
 	// 签署人的手机号，11位数字
 	ApproverMobile *string `json:"ApproverMobile,omitempty" name:"ApproverMobile"`
 
-	// 如果签署方是企业签署方，则为企业名
+	// 如果签署方是企业签署方(approverType = 1 或者 approverType = 3)，
+	// 
+	// 则企业名称必填
 	OrganizationName *string `json:"OrganizationName,omitempty" name:"OrganizationName"`
 
 	// 签署人的签署控件列表
@@ -92,10 +96,14 @@ type ApproverInfo struct {
 	// 合同的强制预览时间：3~300s，未指定则按合同页数计算
 	PreReadTime *int64 `json:"PreReadTime,omitempty" name:"PreReadTime"`
 
-	// 签署人userId，传此字段则不用传姓名、手机号
+	// 签署人userId，仅支持本企业的员工userid， 可在控制台组织管理处获得
+	// 
+	// 若传此字段 则以userid的信息为主，会覆盖传递过来的签署人基本信息， 包括姓名，手机号，证件类型等信息
 	UserId *string `json:"UserId,omitempty" name:"UserId"`
 
-	// 签署人用户来源，企微侧用户请传入：WEWORKAPP
+	// 签署人用户来源，此参数仅针对企微用户开放
+	// 
+	// 企微侧用户请传入：WEWORKAPP
 	ApproverSource *string `json:"ApproverSource,omitempty" name:"ApproverSource"`
 
 	// 企业签署方或签标识，客户自定义，64位长度。用于发起含有或签签署人的合同。或签参与人必须有此字段。合同内不同或签参与人CustomApproverTag需要保证唯一。如果或签签署人为本方企微参与人，ApproverSource参数需要指定WEWORKAPP
@@ -6163,6 +6171,8 @@ type FlowCreateApprover struct {
 	// 自动签署仅进行盖章操作，不能是手写签名。
 	// 本方企业自动签署的签署人会默认是当前的发起人
 	// 他方企业自动签署的签署人是自动签模板的他方企业授权人
+	// 7: 个人自动签署，适用于个人自动签场景。
+	// 注: 个人自动签场景为白名单功能, 使用前请联系对接的客户经理沟通。
 	ApproverType *int64 `json:"ApproverType,omitempty" name:"ApproverType"`
 
 	// 签署人企业名称
@@ -6209,15 +6219,17 @@ type FlowCreateApprover struct {
 	// 合同的强制预览时间：3~300s，未指定则按合同页数计算
 	PreReadTime *uint64 `json:"PreReadTime,omitempty" name:"PreReadTime"`
 
-	// 签署方经办人的电子签用户ID
-	// <br/>当未指定签署人姓名+手机号的情况下，该字段毕传
+	// 签署人userId，仅支持本企业的员工userid， 可在控制台组织管理处获得
+	// 
+	// 若传此字段 则以userid的信息为主，会覆盖传递过来的签署人基本信息， 包括姓名，手机号，证件类型等信息
 	UserId *string `json:"UserId,omitempty" name:"UserId"`
 
 	// 当前只支持true，默认为true
 	Required *bool `json:"Required,omitempty" name:"Required"`
 
-	// 签署人用户来源
-	// <br/>企微侧用户请传入：WEWORKAPP
+	// 签署人用户来源，此参数仅针对企微用户开放
+	// 
+	// 企微侧用户请传入：WEWORKAPP
 	ApproverSource *string `json:"ApproverSource,omitempty" name:"ApproverSource"`
 
 	// 企业签署方或签标识，客户自定义，64位长度
@@ -6248,11 +6260,11 @@ type FlowCreateApprover struct {
 	// <br/>为个人签署方时则由发起方企业审核。
 	ApproverNeedSignReview *bool `json:"ApproverNeedSignReview,omitempty" name:"ApproverNeedSignReview"`
 
-	// 签署人签署控件
+	// 签署人签署控件， 此参数仅针对文件发起（CreateFlowByFiles）生效
 	// <br/>文件发起时，可通过该参数为签署人指定签署控件类型以及位置
 	SignComponents []*Component `json:"SignComponents,omitempty" name:"SignComponents"`
 
-	// 签署人填写控件
+	// 签署人填写控件 此参数仅针对文件发起（CreateFlowByFiles）生效
 	// <br/>文件发起时，可通过该参数为签署人指定填写控件类型以及位置
 	Components []*Component `json:"Components,omitempty" name:"Components"`