+++ tencentcloud/gme/v20180711/models.go
@@ -2467,7 +2467,7 @@ type UpdateScanRoomsRequestParams struct {
 	// 需要送检的所有房间号。多个房间号之间用","分隔。示例："0001,0002,0003"
 	RoomIdString *string `json:"RoomIdString,omitempty" name:"RoomIdString"`
 
-	// 符合此正则表达式规则的房间号将被送检。示例：["^6.*"] 表示所有以6开头的房间号将被送检
+	// 符合此正则表达式规则的房间号将被送检。示例：^6.*（表示所有以6开头的房间号将被送检）
 	RoomIdRegex []*string `json:"RoomIdRegex,omitempty" name:"RoomIdRegex"`
 }
 
@@ -2480,7 +2480,7 @@ type UpdateScanRoomsRequest struct {
 	// 需要送检的所有房间号。多个房间号之间用","分隔。示例："0001,0002,0003"
 	RoomIdString *string `json:"RoomIdString,omitempty" name:"RoomIdString"`
 
-	// 符合此正则表达式规则的房间号将被送检。示例：["^6.*"] 表示所有以6开头的房间号将被送检
+	// 符合此正则表达式规则的房间号将被送检。示例：^6.*（表示所有以6开头的房间号将被送检）
 	RoomIdRegex []*string `json:"RoomIdRegex,omitempty" name:"RoomIdRegex"`
 }