+++ tencentcloud/live/v20180801/models.go
@@ -784,9 +784,9 @@ type CommonMixLayoutParams struct {
 
 	// 当InputType为3(画布)时，该值表示画布的颜色。
 	// 常用的颜色有：
-	// 红色：0xcc0033。
-	// 黄色：0xcc9900。
-	// 绿色：0xcccc33。
+	// 红色：0xCC0033。
+	// 黄色：0xCC9900。
+	// 绿色：0xCCCC33。
 	// 蓝色：0x99CCFF。
 	// 黑色：0x000000。
 	// 白色：0xFFFFFF。