+++ tencentcloud/mps/v20190612/models.go
@@ -14592,7 +14592,7 @@ type VideoTemplateInfo struct {
 	// 注意：av1 编码容器目前只支持 mp4 。
 	Codec *string `json:"Codec,omitempty" name:"Codec"`
 
-	// 视频帧率，取值范围：[0, 100]，单位：Hz。
+	// 视频帧率，取值范围：[0, 120]，单位：Hz。
 	// 当取值为 0，表示帧率和原始视频保持一致。
 	// 注意：自适应码率时取值范围是 [0, 60]
 	Fps *uint64 `json:"Fps,omitempty" name:"Fps"`