+++ tencentcloud/vod/v20180717/models.go
@@ -9831,6 +9831,12 @@ type DescribeMediaProcessUsageDataRequestParams struct {
 	// <li> ExtractTraceWatermark: 提取水印</li>
 	// <li> AddTraceWatermark: 添加水印</li>
 	// <li> RebuildMedia: 音画质重生</li>
+	// <li> QualityInspect: 音画质检测</li>
+	// <li> VideoHighlight: 视频智能集锦</li>
+	// <li> VideoTag: 视频智能标签</li>
+	// <li> VideoClassification:  视频智能分类</li>
+	// <li> VideoCover: 视频智能封面</li>
+	// <li> VideoSegment: 视频智能拆条</li>
 	// <li>Transcode: 转码，包含普通转码、极速高清和视频编辑（不推荐使用）</li>
 	Type *string `json:"Type,omitempty" name:"Type"`
 }
@@ -9859,6 +9865,12 @@ type DescribeMediaProcessUsageDataRequest struct {
 	// <li> ExtractTraceWatermark: 提取水印</li>
 	// <li> AddTraceWatermark: 添加水印</li>
 	// <li> RebuildMedia: 音画质重生</li>
+	// <li> QualityInspect: 音画质检测</li>
+	// <li> VideoHighlight: 视频智能集锦</li>
+	// <li> VideoTag: 视频智能标签</li>
+	// <li> VideoClassification:  视频智能分类</li>
+	// <li> VideoCover: 视频智能封面</li>
+	// <li> VideoSegment: 视频智能拆条</li>
 	// <li>Transcode: 转码，包含普通转码、极速高清和视频编辑（不推荐使用）</li>
 	Type *string `json:"Type,omitempty" name:"Type"`
 }
@@ -19453,6 +19465,7 @@ type ProductInstanceRecource struct {
 	// <li>Transcode：普通转码资源包。</li>
 	// <li>TESHD：极速高清转码资源包。</li>
 	// <li>Review：音视频审核转码资源包。</li>
+	// <li>MediaProcess：媒体处理时长资源包。</li>
 	ResourceType *string `json:"ResourceType,omitempty" name:"ResourceType"`
 
 	// 资源包额度。
@@ -19461,6 +19474,7 @@ type ProductInstanceRecource struct {
 	// <li>音视频审核资源包，单位为秒。</li>
 	// <li>音视频极速高清资源包，单位为秒。</li>
 	// <li>音视频加速资源包，单位为字节。</li>
+	// <li>媒体处理时长资源包，单位为秒。</li>
 	Amount *int64 `json:"Amount,omitempty" name:"Amount"`
 
 	// 资源包余量。
@@ -19469,6 +19483,7 @@ type ProductInstanceRecource struct {
 	// <li>音视频审核资源包，单位为秒。</li>
 	// <li>音视频极速高清资源包，单位为秒。</li>
 	// <li>音视频加速资源包，单位为字节。</li>
+	// <li>媒体处理时长资源包，单位为秒。</li>
 	Left *int64 `json:"Left,omitempty" name:"Left"`
 }
 
@@ -23166,6 +23181,7 @@ type TaskStatData struct {
 	// <li> ExtractTraceWatermark: 提取水印</li>
 	// <li> AddTraceWatermark: 添加水印</li>
 	// <li> RebuildMedia: 音画质重生</li>
+	// <li> QualityInspect: 音画质检测</li>
 	// <li>Transcode: 转码，包含普通转码、极速高清和视频编辑（不推荐使用）</li>
 	TaskType *string `json:"TaskType,omitempty" name:"TaskType"`